+++ src/main/java/dbservice/DAOs/LineDAO.java
@@ -11,22 +11,22 @@ public class LineDAO {
 
 	private final JdbcTemplate jdbcTemplate;
 
-	private Long acquireID() {
+	private Integer acquireID() {
 		String sql = "SELECT TOP 1 line_id FROM Lines ORDER BY line_id DESC";
-		return jdbcTemplate.queryForObject(sql, Long.class);
+		return jdbcTemplate.queryForObject(sql, Integer.class);
 	}
 
 	public LineDAO(JdbcTemplate jdbcTemplate) {
 		this.jdbcTemplate = jdbcTemplate;
 	}
 
-	public Long addLineGetID(Line line) {
+	public Integer addLineGetID(Line line) {
 		String sql = "INSERT INTO Lines (name, first_stop_id) VALUES (?, ?)";
 		jdbcTemplate.update(sql, line.getName(), line.getFirst_stop_id());
 		return acquireID();
 	}
 
-	public Line getLineByID(Long lineID) {
+	public Line getLineByID(Integer lineID) {
 		String sql = "SELECT * FROM Lines WHERE line_id = ?";
 		return jdbcTemplate.queryForObject(sql, new Object[] { lineID }, new BeanPropertyRowMapper<>(Line.class));
 	}
