+++ src/main/java/dbservice/DAOs/TicketDAO.java
@@ -17,29 +17,29 @@ public class TicketDAO {
 		this.jdbcTemplate = jdbcTemplate;
 	}
 
-	private Long acquireID() {
+	private Integer acquireID() {
 		String sql = "SELECT TOP 1 ticket_id FROM Tickets ORDER BY ticket_id DESC";
-		return jdbcTemplate.queryForObject(sql, Long.class);
+		return jdbcTemplate.queryForObject(sql, Integer.class);
 	}
 
-	public Long addTicketGetID(Ticket ticket) {
+	public Integer addTicketGetID(Ticket ticket) {
 		String sql = "INSERT INTO Tickets (traveler_id, first_stop, last_stop, train_id, wagon_num) VALUES (?, ?, ?, ?, ?)";
 		jdbcTemplate.update(sql, ticket.getTravelerID(), ticket.getFirstStop(), ticket.getLastStop(),
 				ticket.getTrainID(), ticket.getWagonNum());
 		return acquireID();
 	}
 
-	public void expireTicket(Long id) {
+	public void expireTicket(Integer id) {
 		String sql = "UPDATE Tickets SET expired = TRUE WHERE ticket_id = ?";
 		jdbcTemplate.update(sql, id);
 	}
 
-	public void unexpireTicket(Long id) {
+	public void unexpireTicket(Integer id) {
 		String sql = "UPDATE Tickets SET expired = FALSE WHERE ticket_id = ?";
 		jdbcTemplate.update(sql, id);
 	}
 
-	public List<Ticket> getTicketsByTravelerID(Long travelerID) {
+	public List<Ticket> getTicketsByTravelerID(Integer travelerID) {
 		String sql = "SELECT * FROM Tickets WHERE traveler_id = ?";
 		return jdbcTemplate.query(sql, BeanPropertyRowMapper.newInstance(Ticket.class), travelerID);
 	}
